requirements.txt
wandb>=0.15.0
torch>=2.0.0
torchvision>=0.15.0
libtmux>=0.15.0
pyyaml>=6.0
pytest>=7.0.0 
prompt_toolkit>=3.0.0